fix(back-end): corrected totalAmount prop passed when new cart is created, transfer decoder in service
